+++ bench/trunc_prec_schunk.c
@@ -48,7 +48,7 @@ int main() {
   int nchunk, nchunks = 0;
   blosc_timestamp_t last, current;
   double totaltime;
-  float totalsize = isize * NCHUNKS;
+  float totalsize = (float)(isize * NCHUNKS);
   double *data_buffer = malloc(CHUNKSIZE * sizeof(double));
   double *rec_buffer = malloc(CHUNKSIZE * sizeof(double));
 
@@ -108,7 +108,7 @@ int main() {
   }
   blosc_set_timestamp(&current);
   totaltime = blosc_elapsed_secs(last, current);
-  totalsize = isize * nchunks;
+  totalsize = (float)(isize * nchunks);
   printf("[Decompr] Elapsed time:\t %6.3f s."
                  "  Processed data: %.3f GB (%.3f GB/s)\n",
          totaltime, totalsize / GB, totalsize / (GB * totaltime));
