+++ examples/noinit.c
@@ -43,7 +43,7 @@ int main(){
   int i;
 
   for(i=0; i<SIZE; i++){
-    data[i] = i;
+    data[i] = (float)i;
   }
 
   /* Register the filter with the library */
