+++ examples/simple.c
@@ -38,7 +38,7 @@ int main() {
   int i, ret;
 
   for (i = 0; i < SIZE; i++) {
-    data[i] = i;
+    data[i] = (float)i;
   }
 
   printf("Blosc version info: %s (%s)\n",
